Fix: assertion error when popping route in the middle of transition animation (#16)
### Issue

Fixes #12, where the following assertion error is thrown when popping a
route during a push transition animation.

```
_AssertionError ('package:navigator_resizable/src/navigator_event_observer.dart': Failed assertion: line 290 pos 12: 'currentRoute.animation!.status == AnimationStatus.reverse': is not true.)
```

The problem was that the
`NavigatorEventObserverState._didPopNextInternal` method, which is
called when a route is popped, assumed that the popped route was always
the one placed on top of the navigation stack before the transition
began (the `currentRoute` in the code). However, this is not always the
case. For example, when `Navigator.pop` is called immediately after
`Navigator.push`ing a route, the popped route is not the previous
top-most route, but rather the route that was pushed by `Navigator.push`
right before. Since this behavior contradicts `_didPopNextInternal`'s
assumption, the assertion error is thrown.

### Solution

The new version of `_didPopNextInternal` makes no assumptions about
where the popped route comes from. Although this change fixes the
assertion issue, it has resulted in losing the ability to track the
top-most route before a transition starts. Because of this, the
`currentRoute` parameter of `NavigatorEventListener.didStartTransition`
was removed.